Implement base data class for wires
"Wires" are an abstract idea to represent a single bit. This data base-class
starts providing the necessary hierarchy to have specific interpretations of
groups of bits. Eventually, the subclassed types (such as uint) will implement
a function required by data, which will write out the wire in the way a
uint (or any other type) supports.

Also includes generic functions to pack and unpack the bits in the object.