Generate method for backends takes a stream to write to now
This means we can control WHERE we generate to. If nil is passed, then a
string is generated, and that is it. If a writable output file-stream is
passed, then the file is written to. If, for some reason, a network socket
were used, the stream would be written there.

The unit test was reworked to have string-generation be the default, because
tha makes it easier to run the tests.