Evaluate Chil unit tests using Lisp code rather than the Makefile
This way fewer things are done with the Makefile and the shell's expansion,
and more of it done inside the relatively uniform universe of Common Lisp.

This reinforces the use of Makefiles as a simple interface for people to
interact with the project, rather than a real way of working on things.

This code is taken and modified from
https://github.com/AccelerationNet/lisp-unit2#asdf.