Start using autoconf to build Makefile and documentation
We do not use automake to generate a Makefile from a Makefile.am and
Makefile.in. Instead, we just use a hand-written Makefile.in, which has some
substitions made.

Automake is needed in guix.scm, because autoconf depends on some files that
automake provides, even though we do not explicitly need automake.