Note that Calyx is technically an IR, not a high-level HDL
This is a fairly vague mention, because YOU can write Calyx and it will
compile down. The original idea of Calyx was that HLS tools could generate
Calyx to express architecture-specific information in the hardware blocks and
the control structure in the control block. Then the Calyx compiler uses that
to generate a synthesizable HDL.