Add check-it system as dependency, to generate automatic tests
check-it is a property-based testing library. This kind of testing means
the testing system generates random test cases for the item being
tested. If a failing input is found, the test system will progressively
reduce the "size" of the input (for whatever size means) until a minimal
case is found that reproduces the issue.